backend/requirements.txt
fastapi
uvicorn[standard]
dspy
python-multipart
python-dotenv
